multiselection in K3bDataFileView length of K3bDataDoc only calculated when adding or removing items K3bDevice::readToc adds "Track X" as default-trackname to K3bTrack cdrdao-driver is used
svn path=/trunk/kdemultimedia/k3b/; revision=123104

+++ src/audio/k3baudiodoc.cpp
@@ -90,8 +90,8 @@ bool K3bAudioDoc::newDocument()
 
 
 
-int K3bAudioDoc::size() const {
-  int size = 0;
+long K3bAudioDoc::size() const {
+  long size = 0;
   for( K3bAudioTrack* _t = m_tracks->first(); _t; _t = m_tracks->next() ) {
     size += _t->size();
   }
